Refactor { train_net, utils/net }
Summary:
This diff starts a series of diffs to refactor the detectron codebase
so it can be used in an elastic data parallel context (see
https://our.intern.facebook.com/intern/dex/caffe2/elastic-data-parallel-model-for-distributed-training/)
without any forks.

Specifically, this diff does the following:

 - Splits out the `TrainingStats` class so it can be reused / composed
 - Slightly refactor of `initialize_weights_from_file()`
 - Reduces the verbosity of some logs

Hopefully, nothing controversial here :)

Reviewed By: rbgirshick

Differential Revision: D6826820

fbshipit-source-id: fc15209a0ff50e5d09281e36173198c77aa77a12
